python_services/watermark_generator/requirements.txt
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.4
scikit-image==0.21.0
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0